Model updates for assignable workouts.
Goal is for this type of structure to be possible.

Template workout: A workout which can be reused and assigned to
different teams. It is reused by creating an assigned workout.

Assigned workout: A workout placeholder which represents that a template
workout will be performed on a selected date. Primarily used for
grouping the workouts by athlete/team. As of right now no plans on
allowing these to be customized on an individual level.

Erg workout or water workout: A workout performed by an athlete or boat,
with power numbers, times, distances, and heart rates.

Essentially this should make it possible to create groupings, view
schedules at a high level, and dig into each individual athlete's
performance.